Refactoring, Cleanup & Commenting
Edited the project structure to ensure every class is in their appropriate package. Refactored the WeatherAPIMain class to act as a runnable class within the framework to ensure a request is being sent to the API and a response is being received. This has console output too. Commented out ConnectionManager and WeatherAPIMain. Removed redundant imports.